TST: Splitting test_datetime.py into smaller chunks (gh14854)
