+++ pandas/tests/indexes/datetimes/test_ops.py
@@ -1,9 +1,11 @@
+import warnings
 import numpy as np
 from datetime import timedelta
 
 import pandas as pd
 import pandas.tslib as tslib
 import pandas.util.testing as tm
+from pandas.core.common import PerformanceWarning
 from pandas import (DatetimeIndex, PeriodIndex, Series, Timestamp, Timedelta,
                     date_range, TimedeltaIndex, _np_version_under1p10, Index,
                     datetime, Float64Index)
@@ -950,3 +952,122 @@ class TestDateTimeIndexToJulianDate(tm.TestCase):
                         freq='S').to_julian_date()
         self.assertIsInstance(r2, Float64Index)
         tm.assert_index_equal(r1, r2)
+
+
+class TestDatetimeIndex(tm.TestCase):
+    _multiprocess_can_split_ = True
+
+    # GH 10699
+    def test_datetime64_with_DateOffset(self):
+        for klass, assert_func in zip([Series, DatetimeIndex],
+                                      [self.assert_series_equal,
+                                       tm.assert_index_equal]):
+            s = klass(date_range('2000-01-01', '2000-01-31'), name='a')
+            result = s + pd.DateOffset(years=1)
+            result2 = pd.DateOffset(years=1) + s
+            exp = klass(date_range('2001-01-01', '2001-01-31'), name='a')
+            assert_func(result, exp)
+            assert_func(result2, exp)
+
+            result = s - pd.DateOffset(years=1)
+            exp = klass(date_range('1999-01-01', '1999-01-31'), name='a')
+            assert_func(result, exp)
+
+            s = klass([Timestamp('2000-01-15 00:15:00', tz='US/Central'),
+                       pd.Timestamp('2000-02-15', tz='US/Central')], name='a')
+            result = s + pd.offsets.Day()
+            result2 = pd.offsets.Day() + s
+            exp = klass([Timestamp('2000-01-16 00:15:00', tz='US/Central'),
+                         Timestamp('2000-02-16', tz='US/Central')], name='a')
+            assert_func(result, exp)
+            assert_func(result2, exp)
+
+            s = klass([Timestamp('2000-01-15 00:15:00', tz='US/Central'),
+                       pd.Timestamp('2000-02-15', tz='US/Central')], name='a')
+            result = s + pd.offsets.MonthEnd()
+            result2 = pd.offsets.MonthEnd() + s
+            exp = klass([Timestamp('2000-01-31 00:15:00', tz='US/Central'),
+                         Timestamp('2000-02-29', tz='US/Central')], name='a')
+            assert_func(result, exp)
+            assert_func(result2, exp)
+
+            # array of offsets - valid for Series only
+            if klass is Series:
+                with tm.assert_produces_warning(PerformanceWarning):
+                    s = klass([Timestamp('2000-1-1'), Timestamp('2000-2-1')])
+                    result = s + Series([pd.offsets.DateOffset(years=1),
+                                         pd.offsets.MonthEnd()])
+                    exp = klass([Timestamp('2001-1-1'), Timestamp('2000-2-29')
+                                 ])
+                    assert_func(result, exp)
+
+                    # same offset
+                    result = s + Series([pd.offsets.DateOffset(years=1),
+                                         pd.offsets.DateOffset(years=1)])
+                    exp = klass([Timestamp('2001-1-1'), Timestamp('2001-2-1')])
+                    assert_func(result, exp)
+
+            s = klass([Timestamp('2000-01-05 00:15:00'),
+                       Timestamp('2000-01-31 00:23:00'),
+                       Timestamp('2000-01-01'),
+                       Timestamp('2000-03-31'),
+                       Timestamp('2000-02-29'),
+                       Timestamp('2000-12-31'),
+                       Timestamp('2000-05-15'),
+                       Timestamp('2001-06-15')])
+
+            # DateOffset relativedelta fastpath
+            relative_kwargs = [('years', 2), ('months', 5), ('days', 3),
+                               ('hours', 5), ('minutes', 10), ('seconds', 2),
+                               ('microseconds', 5)]
+            for i, kwd in enumerate(relative_kwargs):
+                op = pd.DateOffset(**dict([kwd]))
+                assert_func(klass([x + op for x in s]), s + op)
+                assert_func(klass([x - op for x in s]), s - op)
+                op = pd.DateOffset(**dict(relative_kwargs[:i + 1]))
+                assert_func(klass([x + op for x in s]), s + op)
+                assert_func(klass([x - op for x in s]), s - op)
+
+            # assert these are equal on a piecewise basis
+            offsets = ['YearBegin', ('YearBegin', {'month': 5}), 'YearEnd',
+                       ('YearEnd', {'month': 5}), 'MonthBegin', 'MonthEnd',
+                       'SemiMonthEnd', 'SemiMonthBegin',
+                       'Week', ('Week', {
+                           'weekday': 3
+                       }), 'BusinessDay', 'BDay', 'QuarterEnd', 'QuarterBegin',
+                       'CustomBusinessDay', 'CDay', 'CBMonthEnd',
+                       'CBMonthBegin', 'BMonthBegin', 'BMonthEnd',
+                       'BusinessHour', 'BYearBegin', 'BYearEnd',
+                       'BQuarterBegin', ('LastWeekOfMonth', {
+                           'weekday': 2
+                       }), ('FY5253Quarter', {'qtr_with_extra_week': 1,
+                                              'startingMonth': 1,
+                                              'weekday': 2,
+                                              'variation': 'nearest'}),
+                       ('FY5253', {'weekday': 0,
+                                   'startingMonth': 2,
+                                   'variation':
+                                   'nearest'}), ('WeekOfMonth', {'weekday': 2,
+                                                                 'week': 2}),
+                       'Easter', ('DateOffset', {'day': 4}),
+                       ('DateOffset', {'month': 5})]
+
+            with warnings.catch_warnings(record=True):
+                for normalize in (True, False):
+                    for do in offsets:
+                        if isinstance(do, tuple):
+                            do, kwargs = do
+                        else:
+                            do = do
+                            kwargs = {}
+
+                        for n in [0, 5]:
+                            if (do in ['WeekOfMonth', 'LastWeekOfMonth',
+                                       'FY5253Quarter', 'FY5253'] and n == 0):
+                                continue
+                            op = getattr(pd.offsets, do)(n,
+                                                         normalize=normalize,
+                                                         **kwargs)
+                            assert_func(klass([x + op for x in s]), s + op)
+                            assert_func(klass([x - op for x in s]), s - op)
+                            assert_func(klass([op + x for x in s]), op + s)
